+++ src/cache/LocalForageCacheBackend.ts
@@ -3,7 +3,6 @@ import type { CacheBackend } from "./CacheBackend";
 export default class LocalForageCacheBackend implements CacheBackend {
   private readonly store: any;
 
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   constructor(localforage: any) {
     this.store = localforage.createInstance({
       name: "requestCache",